+++ fishfish/ws/models/__init__.py
@@ -1,6 +1,2 @@
-from .domain_delete import WSDomainDelete
-from .domain_create import WSDomainCreate
-from .domain_update import WSDomainUpdate
-from .url_delete import WSUrlDelete
-from .url_create import WSUrlCreate
-from .url_update import WSUrlUpdate
+from .urls import WSUrlDelete, WSUrlCreate, WSUrlUpdate
+from .domains import WSDomainCreate, WSDomainUpdate, WSDomainDelete
